qr scanner layout improvised to fix cont. scanning issue
